review comment

+++ onnxruntime/contrib_ops/cpu/transformers/generation_shared.h
@@ -46,7 +46,7 @@ struct IBeamSearchCpuState {
 
 template <typename T>
 struct IGreedySearchState {
-  gsl::span<int32_t> next_tokens_cpu;   // shape (batch_size)
+  gsl::span<int64_t> next_tokens_cpu;   // shape (batch_size)
   gsl::span<int32_t> sequences_space;   // shape (2, batch_size, max_length)
   gsl::span<int32_t> sequence_lengths;  // shape (batch_size)
   gsl::span<int32_t> next_positions;    // shape (batch_size, num_beams). Next position value for position_ids.
